+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 - Changes recycle/delete to not push useless undo actions
 - Fixes copy names of folders to not use file extension logic
 - Fixes input text selection being reset to start when navigating search history with search strings of the same length
+- Fixes loss of keyboard focus
 
 ### 0.7.1  (2024-11-25)
 - Upgrades to .NET Framework 4.8.1